Reimplement `Bit` and `Register` with more type-safe logic
This is a major rewrite of the `Register` system in particular, but
significantly touches `Bit` as well.

This moves the type safety all the way down into the `BitInfo` and
`RegisterInfo` classes, with the exception that Rust space can
_technically_ have minor data incoherence between `AncillaRegister` and
contained qubits, though the internals are private and the constructors
will not allow this.

As a direct result (and as a deliberate aim), `Bit` and `Register` can
no longer be instantiated directly; there's no real concept of the
typeless base classes in Rust, so it takes quite a lot of gymnastics to
keep the old Python form working (where the old Python code wasn't
necessarily the best use of subclassing already).

The dynamic `Iterator` methods are now statically dispatched instead.
Much of the `inner_X` logic is gone; the classes interact a bit more
cleanly with each other.  The bit and register types are constructed
simultaneously by macro, as are their Python-space variants; this
represents how coupled the classes are, and reduces the amount of
`pub(crate)` interfaces needed to split the files.

Changes in the tests are to account for some standard Python errors
changing from `CircuitError` to the more standard `TypeError`, and to
deal with the fact that `Bit` and `Register` cannot be constructed
directly.  Also, some tests inserting insane and undocumented behaviour
like having `QuantumRegister(2.0)` be valid, are gone.

The `__repr__` of `Bit` is now shorter and a bit more legible.