Adjust drawer register labels
The new repr() implementation for Qubit and QuantumRegister is a better
formatting but it is not valid math mode for drawing. There was an edge
case in the drawer if the circuit has a layout set but the registers in
the virtual qubits of the layout are not present in the layout's
registers it would use the repr for the label. This was previously
something like `Qubit(QuantumRegister(2, "name"))` which while overly
verbose would render. However the intent here is to have a human
readable name so instead this adjusts the logic to just pull the
register name and index from the private register field in the Qubit
object to create that label and if there is no register just use the
index for the layout mapping.