Preserve bit locations through pickle
Previously when we were pickling a DAGCircuit the bit locations fields
were forgotten about. So when we loaded a DAGCircuit from a pickle the
bit locations were empty. This would cause any call to find_bit() to
raise an error because there was no entry for any of the bits in the
dag. This commit fixes this by reconstructing by including the bit
locations fields in the object state returned by __getstate__ and
populating the fields from the provided state in __setstate__/

Fixes #13976