Remove Python dependency from circuit construction
In preparation for providing a C API for building and interacting with
circuits this commit removes the required Python component from the
circuit construction APIs on CircuitData. After #13860 the only use case
of Python in the circuit constructor is to manipulate a
ParameterExpression global phase as ParameterExpression is still a
Python only construct. This moves to using Python::with_gil() in those
places only when it's needed. This gives us a path for building a
circuit without the Python interpreter for the C API.