Add missing  `__getitem__` and other attributes to `BitLocations` (#13997)
* Add: `__getitem__` and other attributes to `BitLocations`
In an oversight after #13860, when switching to use the pyo3 friendly `BitLocations` which is not an instance of `namedtuple` we did not extend it to include sequence methods `__getitem__` and `__len__`, which results in breaking api changes.

The following commits add those methods to `BitLocations` and also labels the class correctly as a `sequence` in PyO3.

* Fix: Address review comments

* Fix: More formatting issues in tests

* Apply suggestions from code review

Co-authored-by: Matthew Treinish <mtreinish@kortar.org>

---------

Co-authored-by: Matthew Treinish <mtreinish@kortar.org>
(cherry picked from commit 3a80a814379f3888fbe26512f07c2ae6a601d2a9)