Preserve bit locations through pickle (#13980)
* Preserve bit locations through pickle

Previously when we were pickling a DAGCircuit the bit locations fields
were forgotten about. So when we loaded a DAGCircuit from a pickle the
bit locations were empty. This would cause any call to find_bit() to
raise an error because there was no entry for any of the bits in the
dag. This commit fixes this by reconstructing the bit locations on
__setstate__ by iterating over the dag's qubits and clbits.

Fixes #13976

* Fix clippy

* Update docstrings

* Fix name change after rebase