Try building aarch64 wheels with Rust 1.83
This commit attempts to build the aarch64 linux wheels with rust 1.83.
When we made aarch64 linux a tier 1 supported platform #13737 there
were packaging issues with the official distrubtions for Rust 1.85 and
1.84 on the platform which were preventing PGO builds from succeeding.
At the time we we determined that our MSRV of 1.79 was able to build
successfully. However, during the 2.0.0rc1 pre-release the wheel build
job failed because of reported LLVM bug when combining LTO and PGO which
we do for the release jobs. Local debugging determined these errors
should be been fixed if built with Rust 1.85. Since we know we're not
able to build with Rust 1.85 or 1.84 this commit tries building with
1.83 to see if this can work. If this does not work, follow up commits
will walk backwards to see if 1.82, 1.81, or 1.80 can build successfully
with LTO+PGO.

Related to: #13983