Fix PGO build
PGO fails when we build qiskit-cext both as cdylib in addition to rlib,
due to missing Python symbols. This could be due to us not setting the
pyo3/extension-module correctly; which we want enabled when we compile
the crate for use in the Python extension crate qiskit-pyext, but
don't want for the standalone cdylib usage.

This PR changes to only building the rlib per default, meaning that
building the qiskit-pyext crate (and running PGO) will not build the
cdylib. To build the cdylib for C standalone mode, one has to request
that crate type via cargo rustc .... --crate-type cdylib. We're
investigating further solutions but this now fixes PGO and the C
standalone tests pass as well.