Allow circuits using only SI units to pass through TimeUnitConversion
In f0576fa64, the `TimeUnitConversion` pass was modified so that it
raises an exception if a circuit has only SI unit delays with no `dt`
value set, but the intent seemed to be catch cases where SI units were
used alongside instructions with `dt` units. Here the conditions for
raising an exception are modified to allow circuits with all SI unit
delays (which usually only happens when working with a backend like
`AerSimulator`) to pass through.

Closes https://github.com/Qiskit/qiskit/issues/14016