Run C tests on all tier 1 platforms
This commit expands the testing matrix for the C API tests to cover all
the supported tier 1 platforms. As tier 1 platforms we document that we
test them all on every PR, historically that's just been the Python
testing, but as we grow the capabilities of our C API in Qiskit we
should expand that guarantee to testing on all the tier 1 platforms too.

There will need to be some iteration done to ensure they all work as
expected, it's unlikely that this will work by default on Windows and we
can tweak the configuration based on the results of CI runs.