Add ctest execution to coverage reporting
This commit updates the coverage CI job to also include execution of the
C API tests. Right now the cext crate is showing 0 coverage because we
only test it without instrumention and don't collect any coverage
numbers for the C API testing. This commit adds a ctest execution to the
coverage job to provide coverage numbers for our C API testing.