Normalize internal scoring on typed mapping
This commit fixes several issues all stemming from the same root cause.
There are multiple mappings of indices to qubits to keep track of and
the previous commits in this PR branch were not leveraging the type
system to keep these sane. This resulted in getting the mapping between
representations incorrect at various points that was causing failures
and/or panics when the mappings were not resolved correctly. This commit
fixes the issue by moving to a typed representation of the layout as
soon as posible. This forces all the places that reference a potential
layout to reason about the various mappings correctly.