Use average error map instead of repeated target querying
The earlier approaches were querying the target for each gate while
iterating over all the trials. This ends up doing far more hashmap
lookups (2x for each gate on a node or edge for each layout) this ends
up dominating the runtime. This also fixes an inconsistency in how the
scoring worked because we weren't correctly accounting for the number of
interactions in the circuit.