Reverse free qubit sort order to use lowest error first
This commit fixes a bug that slipped in during an earlier fix where the
sort order for the free qubits mapping was accidentally flipped. The
intent is to put the lowest error free qubit at the end of the vec so
it's the first popped from the stack when mapping it to a qubit with
solely single qubit operations. However the comparison got reversed
accidentally and that was causing the highest error rate qubits to be
used instead. This commit fixes this oversight and fixes the one fialing
visualization test that was using the wrong qubit compared to the
expected one.