Remove Python dependency for creating a blank dag
This commit removes the py token argument for the dag constructor which
wasn't strictly needed. The only pieces of the dagcircuit struct which
need python is the metadata atribute (which is an `Option<PyDict>`) and
the var tracking sets. This commit rearranges the constructor so it
doesn't actively need python. This is achieved by moving the var
tracking sets inside a once lock which is only initialized in a context
where vars are being used. The metadata is changed to None by default.

The other use case was the control flow module
which wasn't strictly required and is removed in this PR and using the
import once cell mechanism that we use for this everywhere else.