Update security policy for 1.x and 2.x
With the 1.x release we're extending security support for Qiskit to 1
yr. We still only support 1.x for general bugfixes for 6 months after
the 2.0.0 release. But for 1.x if any security vulnerabilities are
identified we will provide fixes for that up to 1 yr after the release
of 2.0.0. This was reflected in the 1.4.0 release notes and is on the
version strategy docs:

https://docs.quantum.ibm.com/open-source/qiskit-sdk-version-strategy

but we forgot to update the security policy document in the Qiskit repo.
This commit fixes this oversight.