Add split path for debug vs release mode in Makefile
This commit updates the makefile organization to add a debug mode path
and a release mode path. The debug mode path is needed for running the
ctests, while release mode is what we want to use for building the clib
for direct use. With this the makefile is able to build in either
release mode or debug mode and the targets are setup so we default to
release mode unless debug is explicitly needed.