Try removing restriction on MSRV for aarch64 PGO builds
With the recently identified manylinux image fixes around some of the
gcc toolsets that were included in the latest cibuildwheel bugfix there
is a chance that this could fix the issues we were facing on aarch64 PGO
builds. This commit attempts to remove the restriction we have to build
with MSRV for PGO in the wheel builds on aarch64 linux to validate
whether this was actually fixed or not.