Split DAGCircuit::compose() into Rust and python
The DAGCircuit::compose() method was previously implemented using most
of the logic via python. Basically it was all python just implemented
via PyO3 calls. This limited the performance of the function, but also
potentially more importantly meant that we didn't have an internal
Rust mechanism for running dag circuit composition. This commit seeks to
address this by splitting out the Python interface from the rust
implementation and rewriting the internals to operate on the Rust data
model directly instead of going through python. The `DAGCircuit::compose()`
method becomes the rust function for doing composition, and
`DAGCircuit::py_compose()` is the python interface.

Ideally we'd be able to drop the Python token argument from the new
`DAGCircuit::compose()` too, but because of
`DAGCircuit::apply_operation_back()` and handling of Vars this wasn't
worth the effort of using `Python::with_gil()` since we'd
unconditionally need the gil. Once `apply_operation_back()` no longer
needs the GIL, we should be able to drop the py argument from compose
too.