Port disjoint layout utils in rust
This commit ports the disjoint layout utils into rust. The eventual goal
of this is to enable porting the full DenseLayout and SabreLayout pass
to rust so it's a standalone function. This will enable us exposing
these passes as a C function. Since the eventual goal of this
functionality is to expose it as a C function this attempts to do as
much as possible without a py token. A standalone rust entrypoint is
added which will serve as what is called from the standalone rust passes
in the future, this is then wrapped in a python entrypoint.

If we have a target that can't build a coupling map from Rust the
disjoint utils return None. This could be because the target could be
global and not have any connectivity constraints, or it could be caused
by having a FakeTarget that doesn't expose certain constraints to rust.
This commit fixes this by falling back to the coupling map python path
in these cases since we can get the python coupling map and use the
python version of the disjoint utils.