Port disjoint layout utils in rust
This commit ports the disjoint layout utils into rust. The eventual goal
of this is to enable porting the full DenseLayout and SabreLayout pass
to rust so it's a standalone function. This will enable us exposing
these passes as a C function. Since the eventual goal of this
functionality is to expose it as a C function this attempts to do as
much as possible without a py token. A standalone rust entrypoint is
added which will serve as what is called from the standalone rust passes
in the future, this is then wrapped in a python entrypoint.

One minor note is the visibility of the run pass over components
entrypoint is `pub(crate)`, but the intent is for it to be `pub`. This
visibility was limited here to avoid changing the target's visbility
level (which is `pub(crate)` for some reason). This is already being
fixed in #14056 and when that merges we can update this PR.