Add missing argument list to quantum_volume() docs (#14158)
This commit adds the missing description of the function's arguments to
the quantum_volume(). The docstring was largely copied from the
QuantumVolume class's docstring, but because it was a class the
arguments were documented in the `__init__()` method and were missed in
the function's docstring. This commit corrects that oversight.

(cherry picked from commit 24038aa2aaee6391917a5e157e6f7df00c7a26e7)