Improve annotated plugin (#13916)
* Moving all internal fields into a separate class

* making all the inner functions to be global

* moving handling annotated operations to plugin

The interface is now quite ugly, and we also have an extra isinstance check for
each gate, we will see if it can be removed.

* improving control-modified synthesis plugin

* Changing internal functions to work with QuantumCircuits instead of DAGCircuits

* renaming

* Removing unnecessary argument use_ancillas

This argument is not needed as the qubit tracker already restricts
allowed ancilla qubits.

* progress

* minor

* removing context

* minor

* removing unused argument

* removing an obsolete statement

* minor cleanup

* minor

* improvements to run method

* cleanup

* another pass over the run method + black

* minor cleanup

* pass over synthesize_operation

* more cleanup

* pass over HLS::run

* cleanup

* pass over annotated plugin

* cleanup

* improving comment

* fixing pylint

* remove print statements in tests

* fmt

* fix + test for controlling circuits with nontrivial phase

* adding release notes

* adding test function docstring

* minor

* Finalizing data class contruction

* adding error message

* Adding method op_names to HighLevelSynthesisPluginManager

This is used to simplify the check whether a given node/operation can be skipped.

* updating old usages of equiv library

* fixing lint errors after merge

* slightly simplifying the arguments to instruction_supported method

* Fixing definition method for PyGate and PyInstruction

* docstring fixes

* applying Julien's suggestion

* adding intern

* replacing data by _data (accidentally removed in cleanup)

* porting main functionality to rust

* Adding missing functionality

* passing empty dict to dag_to_circuit when None

* Adding definition for UnitaryGate

Even when HighLevelSynthesis runs directly after UnitarySynthesis, it may happen that
certain definitions involve UnitaryGates (for instance, this is the case for Isometry),
in which case the default behavior of HighLevelSynthesis should be to query the
definition of UnitaryGates if they are not in the basis.

* Replacing from_circuit_data by clone_empty_like

Going from QuantumCircuit/DAGCircuit to CircuitData discards information,
such as information about registers, input variables and more. The previously
tried approach of constructing DAGCircuit/QuantumCircuitData in rust space by
calling _from_circuit_data and manually fixing registers was not complete.
Instead we now call clone_empty_like both on DAGCircuit and QuantumCircuit.

* making target of type Option<Py<Target>>
and avoiding expensive cloning

* changing type of equivalence library as well

* adding pickling support

this actually requires dill since HLSConfig includes a lambda function
for comparing circuits obtained with different synthesis plugins,
but fortunately it's all already supported

* Passing Bound<HighLevelSynthesisData> to functions

This avoids cloning data when calling Python

* Using Bound<QubitTracker> in functions

This avoid having to clone it when calling the Python space

* Removing the pub keywork in front of HLS structs, members and internal functions

* Adding function to return a definition for a given operation.

The main functionality is to create some default definition for unitary gates, for all other
operation types we can simply use the definition method.

* pylint

* clippy

* Adding tests for HLS tracking global phase.

With various conversions of DAGCircuit to QuantumCircuitData to CircuitData in Rust,
it's best to make sure that the global phases appearing in circuits/control-flow subcircuits
are tracked correctly

* cleanup

* removing unnecessary changes to other files

* improving Rust HLS interface; skipping synthesis of already supported ops when calls from annotated plugin

* Moving the default plugin for synthesizing annotated operations to a separate file

* Fixing synthesis of an annotated operation with an empty list of modifiers; adding test

* Recursively combining the modifiers and canonicalizing this list.

This leads to significantly improved synthesis results when adding multiple controls to a gate. Previously,
first controlling a CX with 2 controls, and then controlling the results with 3 controls, first resulted in expanding
the CCCX gates into single- and two-qubits gates, adding controls to each of these gates, and then expanding each
of these controlled gates. Now, a single MCX gate will be produced, which will then be expanded using the best MCX
synthesis algorithm available.

* Adding tests that verify that HLS correctly combines control states for control modifiers

* improving handling of CZ-gates; adding quality tests for controlled X and Z gates

* clippy

* clippy

* Simple version of conjugate reduction.

This is a simplified version from the unused code from the OptimizeAnnotated transpiler pass.
It's especially useful for controlled QFT-adder and similar circuits.

* Adding annotated argument to adder_qft_d00

This allows to let the default QFT-based adder plugins to create the inverse QFTGate as an
AnnotatedOperation, allowing to exploit optimizations in the HighLevelSynthesis transpiler pass

* adjusting HLS plugins + adding quality test for controlled qft adder

* preliminary release notes

* Fully adapting solution for synthesizing base operation of an annotated gates from add_control, but with the twist that some other high-level gate names are also supported; fixing failing tests that are synthesized differently (and actually better) using the new flow

* fix conflicts after merge

* Fixing test to check for gates that should belong to the given multiplier synthesis algorithm, and not the synthesis of the underlying QFT gate.

* Adding comments

* updating code after merging with main

* finalize merge with main

* using pyo3 get/set

* avoding extra casts u32 -> usize when possible

* avoiding multiple calls to data.borrow()

* changing QubitTracker from being a Python class

* applying Eli's review comments

* also applying renames on the python side

* Applying more of Eli's comments

* typo

* Addressing the review comment of creating a single source of truth list for deciding which gates have efficient synthesis algorithms for their controlled versions

* extending tests to control-annotated gates

* review comment

* Improving conjugate reduction and adding tests

* making the test more explicit

* lint

* improving reno

* additional reno

* adding tests for CZGate

* minor cleanup

* typo