Fix cargo clippy warnings in Rust 1.86
In the recently released Rust 1.86 new clippy rules were added/enabled
by default which were causing warnings when using it to run clippy
locally. WHile we run clippy with the msrv in CI to avoid disruptions on
every release these clippy rules typically flag real issues in the code
that would be good to address even if there isn't a CI failure
associated with it. This commit makes the updates flagged by clippy.