Port the full VF2Layout pass to Rust (#14056)
* Port the full vf2layout pass to Rust

This commit provides a full path Rust implementation of the VF2Layout
pass that performs the full pass logic in Rust. We had most of the
pass in Rust already by leveraging rustworkx for its vf2
implementation and the scoring logic in Rust. So all that we did in
Python was create the graphs as inputs to vf2, building the error map
and other rust view objects (since the scoring in Rust predates our
rust data model in Qiskit) and then process each found layout for final
output. This means the performance improvement expected by this commit
is not large since most of the heavy lifting was already in Rust.
Instead what this really enables is running the pass in a Python-free
context, mainly in a C API for executing this pass which will come in a
follow up.

This rust implementation does not cover the full feature set of what the
Python implementation offers. Mainly in the case of no target, coupling
graph randomization (which is a shame we forgot to remove in 2.0, since it
almost always produces worse results), or a user provided error map we
rely on the existing Python implementation. This is a tradeoff because
these use cases are non-standard and having a Rust interface for running
in that mode provides little benefit for the interfaces we're building.

Fixes #12277

* Fix idle qubit handling in layout and free qubit mapping

This commit address some of the panics and failures seen in scheduling
tests. The source of the errors was two fold, the first was the layout
definition coming from the mapping is sparse but we were using nlayout
which doesn't handle that well so we ended up with several sentinel
values persiting to scoring when scoring the layout which caused errors.
The second issue was we were missing the special handling of free qubits
that was added as a performance optimization in #9148 which was causing
other failures since some of the code was expecting that.

* Fix circuit->dag conversion to preserve vars

* Normalize internal scoring on typed mapping

This commit fixes several issues all stemming from the same root cause.
There are multiple mappings of indices to qubits to keep track of and
the previous commits in this PR branch were not leveraging the type
system to keep these sane. This resulted in getting the mapping between
representations incorrect at various points that was causing failures
and/or panics when the mappings were not resolved correctly. This commit
fixes the issue by moving to a typed representation of the layout as
soon as posible. This forces all the places that reference a potential
layout to reason about the various mappings correctly.

* Remove log assertions for rust code path

Now that the pass runs in rust we don't emit python logs for the running
of the pass. This commit removes the test log assertions since they're
no longer present.

* Fix free qubit mapping

This commit fixes the logic in the free qubit mapping and also skips
running vf2 if there are no qubits in the interaction graph and there
are free qubits.

* Make Target public

* Use average error map instead of repeated target querying

The earlier approaches were querying the target for each gate while
iterating over all the trials. This ends up doing far more hashmap
lookups (2x for each gate on a node or edge for each layout) this ends
up dominating the runtime. This also fixes an inconsistency in how the
scoring worked because we weren't correctly accounting for the number of
interactions in the circuit.

* Don't panic in free qubits if there are no error rates

* Reverse free qubit sort order to use lowest error first

This commit fixes a bug that slipped in during an earlier fix where the
sort order for the free qubits mapping was accidentally flipped. The
intent is to put the lowest error free qubit at the end of the vec so
it's the first popped from the stack when mapping it to a qubit with
solely single qubit operations. However the comparison got reversed
accidentally and that was causing the highest error rate qubits to be
used instead. This commit fixes this oversight and fixes the one fialing
visualization test that was using the wrong qubit compared to the
expected one.

* Rename internal scoring function since it doesn't use the target directly anymore

* Stop caching sort key in free qubit mapping

* Fix typ in comment